src/python-engine/requirements.txt
fastapi
uvicorn
browser-use
langchain-openai
langchain-community
python-dotenv
playwright
boto3
